Use community service form elements in widgets
This simplifies them a great deal since logic has been moved to the
elements.
