+++ web/modules/custom/dbcdk_community_reference_field/src/Plugin/Field/FieldWidget/GroupReferenceAutocomplete.php
@@ -2,7 +2,9 @@
 
 namespace Drupal\dbcdk_community_reference_field\Plugin\Field\FieldWidget;
 
-use Symfony\Component\DependencyInjection\ContainerInterface;
+use Drupal\Core\Field\FieldItemListInterface;
+use Drupal\Core\Field\WidgetBase;
+use Drupal\Core\Form\FormStateInterface;
 
 /**
  * Widget for managing references to community service groups.
@@ -17,32 +19,32 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *   }
  * )
  */
-class GroupReferenceAutocomplete extends RemoteReferenceAutocomplete {
+class GroupReferenceAutocomplete extends WidgetBase {
 
   /**
    * {@inheritdoc}
    */
-  public static function create(
-    ContainerInterface $container,
-    array $configuration,
-    $plugin_id,
-    $plugin_definition
-  ) {
-    return new static(
-      $plugin_id,
-      $plugin_definition,
-      $configuration['field_definition'],
-      $configuration['settings'],
-      $configuration['third_party_settings'],
-      $container->get('dbcdk_community_reference_field.id_value_mapper.group')
-    );
+  public static function defaultSettings() {
+    return [] + parent::defaultSettings();
   }
 
   /**
    * {@inheritdoc}
    */
-  protected function getAutocompleteRoute() {
-    return 'dbcdk_community_reference_field.group_autocomplete_controller_autocomplete';
+  public function settingsForm(array $form, FormStateInterface $form_state) {
+    return [];
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function formElement(FieldItemListInterface $items, $delta, array $element, array &$form, FormStateInterface $form_state) {
+    $element += [
+      '#type' => 'dbcdk_community_group_reference_autocomplete',
+      '#default_value' => $items[$delta]->value,
+    ];
+
+    return $element;
   }
 
 }
